Refactoring: rename files, clean files
