--- tests/unit/lib/base.test.ts
@@ -1,7 +1,0 @@
-import { describe, expect, it } from 'vitest'
-
-describe('tests', () => {
-  it('should works', () => {
-    expect(1 + 1).toEqual(2)
-  })
-})
